fix(rdb): fix chunked SBF loading for filters >= 4 GiB

+++ src/server/rdb_load.cc
@@ -1890,13 +1890,13 @@ auto RdbLoaderBase::ReadSBFImpl(bool chunking) -> io::Result<OpaqueObj> {
     SET_OR_UNEXPECT(LoadLen(nullptr), hash_cnt);
 
     if (chunking) {
-      unsigned total_size = 0;
+      size_t total_size = 0;
       SET_OR_UNEXPECT(LoadLen(nullptr), total_size);
 
       filter_data.resize(total_size);
       size_t offset = 0;
       while (offset < total_size) {
-        unsigned chunk_size = 0;
+        size_t chunk_size = 0;
         SET_OR_UNEXPECT(LoadLen(nullptr), chunk_size);
         error_code ec = FetchBuf(chunk_size, filter_data.data() + offset);
         if (ec) {
